requirements.txt
setuptools~=56.0.0
colorama~=0.4.4
pyserial~=3.5
pandas~=1.3.5
matplotlib~=3.5.1